Move randomizing graph and iterator to tests
