+++ jena-core/src/test/java/org/apache/jena/shared/RandomOrderGraph.java
@@ -18,10 +18,14 @@
 
 package org.apache.jena.shared;
 
-import org.apache.jena.graph.* ;
+import org.apache.jena.graph.Graph;
+import org.apache.jena.graph.GraphMemFactory;
+import org.apache.jena.graph.Node;
+import org.apache.jena.graph.Triple;
 import org.apache.jena.graph.impl.WrappedGraph ;
-import org.apache.jena.rdf.model.* ;
-import org.apache.jena.util.iterator.* ;
+import org.apache.jena.rdf.model.Model;
+import org.apache.jena.rdf.model.ModelFactory;
+import org.apache.jena.util.iterator.ExtendedIterator;
 
 /**
  * Wraps a graph and randomizes the order of find results.
